+++ resources/components/laraknife/panels/index-button.blade.php
@@ -1,9 +1,6 @@
 @props(['buttonType' => 'new'])
-<div class="lkn-behind-filter">
-    <div class="row">
+    <div class="row lkn-behind-filter">
         @if ($buttonType === 'new')
         <x-laraknife.buttons.new width1="0" width2="12" />
         @endif
-        {{$slot}}
     </div>
-</div>